Initial insert of SessionLocal
SessionLocal should instantiate a single SQLAlchemy session per request.
To make this work with AiiDA, an AiiDA backend should be sub-classed and
the handling of "get_session" should be updated in order to make sure
QueuePool is not overrun when creating several requests (QueryBuilder
issue).

Also, move exceptions to separate place.