Avoid loading AiiDA Nodes
Retrieving sites and kinds via the QueryBuilder, projecting attributes,
greatly diminishes the need for loading StructureData Nodes.
Further importing general functions from aiida.orm.nodes.data.structures
and recreating the methods under Kinds and StructureData allows for a
complete removal of the use of the AiiDA Node entity in
StructureDataParser.

When saving the OPTiMaDe fields in Node extras, they are saved directly
in the DB using the backend specific API.
This completely avoids ever having to load a Node.

The functionality for loading a Node has been kept, if needed at a later
stage, but at this point a Node will never be loaded.

Co-authored-by: Giovanni Pizzi <giovanni.pizzi@epfl.ch>