Update data_returned
An attempt has been made at reducing AiiDA queries, when determining
data_returned. However, it should be noted that nested lists in the
filters query are not accounted for. This means that for a simple AND
filter query, A AND B is not recognized as being equal to B AND A, and
the data_returned is re-calculated.

This may be implemented, but is at this point considered to be more work
than gain.