Use optimade.__api_version__ instead of CONFIG
Set default page_limit to 15, in order to be below 1 MB in returned
results for structures.

Minor updates to README.md. Most major one is to redirect OPTiMaDe badge
to optimade-python-tools tag (which doesn't exists at this moment).

Update task set-optimade-ver to optimade-req, to be able to set the
requirement on the dependency on `optimade`, i.e., optimade-python-tools