Update middleware use
Instead of using Starlette's middleware functionality, all routers that
use AiiDA are now decorated with the `close_session` decorator.
This is due to the creation of a new thread when the request is handled
in Starlette's middleware. With the decorator, the session is closed
after the _actual_ request has been handled, making sure we're in the
same thread always, retrieving the correct global scoped session and
closing it.

This makes Starlette's middleware superfluous.