Use custom TestClient
Subclass FastAPI/Starlette's `TestClient` in order to prepend path
requests with the version path.
This is done, since the `TestClient` otherwise uses
`urllib.parse.urljoin` to combine the custom `base_url` attribute with
the requested path URI. In doing this, any paths included in the
`base_url` will be removed and only the requested path URI will remaing,
using the scheme and netloc of `base_url`.
So this is a way of injecting the OPTIMADE versioned base URL into the
requests.

A precaution is taken to not override requested complete URLs by
checking for a scheme after using `urllib.parse.urlparse` on the
requested URL.

Furthermore, the client is wrapped in a factory function in order to
create a client with a specific versioned base URL.
If no version is passed to the factory function, the current OPTIMADE
API version's major part it used (the mandatory versioned base URL).