A bunch of minor perk-ups
Use validator in pytests.
For this to work, a client with `raise_server_exceptions=False` had to
be used. This is now called `remote_client`.

Instantiate `LarkParser` with `version=None` to retrieve the latest
version parser automatically.

Subclass `OptimadeWarning` for warnings to ensure all raised warnings
are included as warnings entities in the response.