Update MANIFEST, default to config.json in repo
The MANIFEST.in will now include data JSON files in the source
distribution.
The logging made in the optimade sample server will be done here as well
for announcing which config file is being used, if any.

The CLI will automatically use the repo's config.json if
OPTIMADE_CONFIG_FILE env var is not found or does not point to an
existing file.