Introduce 'aiida-optimade calc' CLI command
The 'calc' command takes in multiple field values, example:

  aiida-optimade calc elements species

In this example the fields `elements` and `species` will be removed from
all Nodes where it exists in the `optimade` Extra.
Then the `_check_and_calculate_entities()` method of the
`AiidaCollection` class will be run with relevant fields set.
This will run the mapper, which will update _only_ the missing fields
for all the Nodes.

Add tests for aiida-optimade calc.

Update some tests for aiida-optimade run according to the new uvicorn
handling of `log-level`.

Add deletion statements of Nodes in memory.
This seems to be an issue especially for the SQLAlchemy backend.