Finish implementing CifDataTranslator
Add test for 5 CifData Nodes when using aiida-optimade init (CLI).

Utilize `node_type` to determine which Translator class to use. For this
purpose, `type` has been aliased as `node_type`, even though that's not
completely true, but this is an easy way to ensure `node_type` is added
to the projection in any given QueryBuilder query.

Minor optimizations have also been implemented:
- Only call `_get_unique_node_property` once per attribute creation
using StructureDataTranslator.
  Since this method uses QueryBuilder, this ensures only a single
QueryBuilder instance is created per attributes creation.
- Delete __node property from Translator classes.

CifDataTranslator will always load in the Node to memory, as this is
needed to utilize the `get_structure` method, which returns a
StructureData Node from a CifData Node, which can be utilized to
calculate the OPTIMADE fields.
This should/will be optimized in the future!

pydantic has become a dependency, since the `get_structure` method uses
pydantic to convert CifData to StructureData.