Introduce initialization with minimization
New `--minimized-fields` option for `aiida-optimade init`.
This will try to avoid calculating big-valued OPTIONAL fields.

To ensure this, all translator field methods have gotten a new
property `store` on whether the calculated field should be stored in the
`optimade` extra or not. This is to be able to calculate some fields
that depend on others (like `structure_features`), but still _only_
saving the value of `structure_features` and not the others.

Specifically for `structure_features` a post-treatment is needed to
ensure the correct features are set/unset.