Add possibility to use a MongoDB
Through the `--mongo` option for `aiida-optimade init` the
AiiDA-OPTIMADE server will be initialized with a MongoDB.
All relevant fields will be added to each document, representing the
relevant StructureData or CifData Node.

The tests have been updated to reflect the new possibility, along with
relevant code in the code base, where needed.

The tests can be run with MongoDB by setting the env var
PYTEST_OPTIMADE_CONFIG_FILE, the value being a path to the config file
to be used.