Remove dependencies gotten from aiida-core (#247)
Since both click, click-completion, and tqdm are core
dependencies of aiida-core (AiiDA), it makes sense to get these
directly from what is supported in AiiDA, instead of trying to
support only newer versions in the current package.

In order to make clear what packages that are dependencies of
aiida-core, but are still used directly in this package, a comment
with a list of these dependencies have been added to
requirements.txt. Should they ever be moved out of AiiDA's core
dependencies they should be added back to this package explicitly.

Included mention of dependencies from optimade in the comment.